Added operator overloading, casting, simpler functions, an example and a Makefile
Overloaded the +, -, * and / operators.
Added a static_cast equivalent, tuple_cast.
Modified the existing functions to simplify the calls. No need to specify the size of the tuple anymore.
A simple example showing the add function, the cast and the overloaded add.
Added a Makefile to help new users build the example.
Added some comments.